+++ src/main/java/calculator/SimpleNumberValidator.java
@@ -1,0 +1,20 @@
+package calculator;
+
+public class SimpleNumberValidator implements NumberValidator {
+
+    private static String DIGIT_REGEX_PATTE = "\\d+";
+
+
+    public void validateNumbers(String[] splittedString) {
+        for (String str : splittedString) {
+            if (!isDigit(str)) {
+                throw new IllegalArgumentException("유효하지 않은 문자 : %s".formatted(str));
+            }
+        }
+    }
+
+    private boolean isDigit(String str) {
+        return str.matches(DIGIT_REGEX_PATTE);
+    }
+
+}
